Refactoring to improve readability and modularity.
