+++ raw_tiles/index/util.py
@@ -1,0 +1,21 @@
+from __future__ import absolute_import
+from itertools import izip
+
+
+def deassoc(x):
+    """
+    Turns an array consisting of alternating key-value pairs into a
+    dictionary.
+
+    Osm2pgsql stores the tags for ways and relations in the planet_osm_ways and
+    planet_osm_rels tables in this format. Hstore would make more sense now,
+    but this encoding pre-dates the common availability of hstore.
+
+    Example:
+    >>> from raw_tiles.index.util import deassoc
+    >>> deassoc(['a', 1, 'b', 'B', 'c', 3.14])
+    {'a': 1, 'c': 3.14, 'b': 'B'}
+    """
+
+    pairs = [iter(x)] * 2
+    return dict(izip(*pairs))
